requirement.txt
# requirements.txt
langchain-anthropic
langchain-openai
python-dotenv
langchain-chroma
langchain
langchain-voyageai